Add an abstraction layer for files
Yafut operates on files using POSIX APIs provided by the operating
system and POSIX-inspired APIs provided by Yaffs code.  Since the tool
is capable of transferring files both from a Yaffs file system to a
host-mounted file system and the other way round, it needs to support a
similar set of operations for both Yaffs files and local files.  Add an
abstraction layer defining an interface for interacting with files of
either of these types to improve code readability and organization.